refactors and styling
