+++ app/assets/stylesheets/article-show.scss
@@ -552,27 +552,65 @@ article {
   }
 }
 
-// Align styles of freezeframed animated images with normal image styles
 .ff-container {
   left: 50%;
   transform: translateX(-50%);
 
+  // Align styles of freezeframed animated images with normal image styles
   canvas {
     border-radius: var(--radius);
   }
 
   img[data-animated='true'] {
     margin: 0;
-    cursor: unset;
+    cursor: pointer;
   }
 
-  .ff-button {
+  .gif-button {
+    --bg: var(--gif-bg);
+    --bg-hover: var(--gif-bg);
+    --color: var(--gif-color);
+    --color-hover: var(--gif-color);
+
     position: absolute;
     bottom: 0;
     right: 0;
     margin: var(--su-1);
-    z-index: 5;
-    background-color: black;
-    color: white;
+    z-index: var(--z-sticky);
+    border-radius: var(--radius);
+    display: flex;
+    align-items: center;
+    padding: var(--su-1) var(--su-2);
+    font-weight: var(--fw-bold);
+
+    &:hover:enabled,
+    .js-focus-visible &.focus-visible:focus,
+    &[aria-pressed='true'] {
+      --bg: var(--gif-bg);
+      --bg-hover: var(--gif-bg);
+      --color: var(--gif-color);
+      --color-hover: var(--gif-color);
+      z-index: var(--z-sticky);
+    }
+  }
+
+  .gif-button[aria-pressed='false'] {
+    .gif-pause {
+      display: inline-block;
+    }
+
+    .gif-play {
+      display: none;
+    }
+  }
+
+  .gif-button[aria-pressed='true'] {
+    .gif-pause {
+      display: none;
+    }
+
+    .gif-play {
+      display: inline-block;
+    }
   }
 }
