+++ app/assets/stylesheets/config/_colors.css
@@ -251,7 +251,7 @@
 
   --tooltip-bg: rgba(var(--grey-900), 0.9);
   --tooltip-color: rgb(var(--white));
-  
+
   /***********************************************
   ** Indicators **********************************
   ***********************************************/
@@ -275,6 +275,13 @@
   /* Indicators: Info */
   --indicator-info-bg: var(--accent-brand);
   --indicator-info-color: rgb(var(--white));
+
+  /***********************************************
+  ** GIFs ****************************************
+  ***********************************************/
+
+  --gif-bg: rgb(var(--black));
+  --gif-color: rgb(var(--white));
 
   /***********************************************
   ** Utilities ***********************************
